+++ sdk/resourcemanager/msi/armmsi/CHANGELOG.md
@@ -1,5 +1,31 @@
 # Release History
 
+## 1.0.0 (2023-02-01)
+### Breaking Changes
+
+- Function `*UserAssignedIdentitiesClient.NewListAssociatedResourcesPager` has been removed
+- Struct `AssociatedResourcesListResult` has been removed
+- Struct `AzureResource` has been removed
+- Struct `CloudError` has been removed
+- Struct `CloudErrorBody` has been removed
+- Struct `UserAssignedIdentitiesClientListAssociatedResourcesResponse` has been removed
+
+### Features Added
+
+- New type alias `CreatedByType` with values `CreatedByTypeApplication`, `CreatedByTypeKey`, `CreatedByTypeManagedIdentity`, `CreatedByTypeUser`
+- New function `timeRFC3339.MarshalText() ([]byte, error)`
+- New function `*timeRFC3339.Parse(string) error`
+- New function `*timeRFC3339.UnmarshalText([]byte) error`
+- New struct `SystemData`
+- New field `SystemData` in struct `FederatedIdentityCredential`
+- New field `SystemData` in struct `Identity`
+- New field `SystemData` in struct `IdentityUpdate`
+- New field `SystemData` in struct `ProxyResource`
+- New field `SystemData` in struct `Resource`
+- New field `SystemData` in struct `SystemAssignedIdentity`
+- New field `SystemData` in struct `TrackedResource`
+
+
 ## 0.7.0 (2022-06-27)
 ### Features Added